backend/requirements.txt
transformers==4.41.2
torch==2.3.0+cpu
torchvision==0.18.0+cpu
torchaudio==2.3.0+cpu
fastapi==0.110.0
uvicorn==0.29.0
pydantic==1.10.13
tqdm
huggingface-hub
motor
python-dotenv
